+++ autolog-core/src/main/java/com/github/maximevw/autolog/core/annotations/AutoLogPerformance.java
@@ -26,6 +26,9 @@ import com.github.maximevw.autolog.core.logger.performance.AdditionalDataProvide
 import net.logstash.logback.argument.StructuredArguments;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.logging.log4j.CloseableThreadContext;
+import org.apache.velocity.Template;
+import org.apache.velocity.app.VelocityEngine;
+import org.apache.velocity.runtime.resource.ResourceManager;
 import org.apiguardian.api.API;
 import org.slf4j.MDC;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -123,6 +126,11 @@ import java.lang.annotation.Target;
 @Retention(RetentionPolicy.RUNTIME)
 @Target({ElementType.METHOD, ElementType.TYPE})
 public @interface AutoLogPerformance {
+
+	/**
+	 * Default Velocity template name used to log the performance information.
+	 */
+	String DEFAULT_VELOCITY_TEMPLATE_NAME = "file:default_performance.vm";
 
 	/**
 	 * @return The name of the performance timer. By default, the name of the invoked method will be used.
@@ -263,4 +271,45 @@ public @interface AutoLogPerformance {
 	 */
 	@API(status = API.Status.STABLE, since = "1.2.0")
 	boolean callerClassAsTopic() default false;
+
+	/**
+	 * @return The Velocity message template that will be used to log the performance information. By default:
+	 *         {@value #DEFAULT_VELOCITY_TEMPLATE_NAME} (this template is provided by the Autolog library).
+	 * <p>
+	 *     The value of this property can be:
+	 *     <ul>
+	 *         <li>the template content itself;</li>
+	 *         <li>an available resource in the application classpath or a path to a file containing the template. In
+	 *         both cases, the template name must be prefixed with "{@code file:}", e.g.
+	 *         {@code file:custom_template.vm}</li>
+	 *     </ul>
+	 * </p>
+	 * <p>
+	 *     The defined template uses Velocity Template Language (see:
+	 *     <a href="https://velocity.apache.org/engine/2.0/vtl-reference.html">Velocity documentation</a>) and can use
+	 *     the following variables:
+	 *     <ul>
+	 *         <li><b>{@code isFailed}</b> ({@code boolean}) indicating if the execution of the invoked method has
+	 *         failed;</li>
+	 *         <li><b>{@code invokedMethod}</b> ({@code String}) containing the name of the invoked method;</li>
+	 *         <li><b>{@code httpMethod}</b> ({@code String}) containing the HTTP method (if applicable);</li>
+	 *         <li><b>{@code executionTime}</b> ({@code String}) containing the execution time of the invoked method in
+	 *         a human-readable format;</li>
+	 *         <li><b>{@code processedItems}</b> ({@code Integer}) containing the number of items processed by the
+	 *         invoked method, only if an {@link AdditionalDataProvider} is provided in the context;</li>
+	 *         <li><b>{@code averageExecutionTimeByItem}</b> ({@code String}) containing the average execution time by
+	 *         processed item in a human-readable format, only if an {@link AdditionalDataProvider} is provided in the
+	 *         context;</li>
+	 *         <li><b>{@code startTime}</b> ({@code String}) containing the start time of the invocation;</li>
+	 *         <li><b>{@code endTime}</b> ({@code String}) containing the end time of the invocation;</li>
+	 *         <li><b>{@code comments}</b> ({@code List<String>}) containing the optional comments, only if an
+	 *         {@link AdditionalDataProvider} is provided in the context.</li>
+	 *     </ul>
+	 * </p>
+	 * @see VelocityEngine
+	 * @see Template
+	 * @see ResourceManager
+	 */
+	@API(status = API.Status.STABLE, since = "1.2.0")
+	String messageTemplate() default DEFAULT_VELOCITY_TEMPLATE_NAME;
 }